Adding OHRRPGCE game engine to the list
OHRRPGCE is developed since 1996. In 2005 the source code is available on GPL license. After more than 20 years, it is still developed by James Paige. Maybe it's not popular, but I think it deserves more love from game developers community. The docs are well written and covers a lot of examples.